feat: clamp for softmax
