+++ atlas-onnx-tracer/src/utils/quantize.rs
@@ -87,10 +87,7 @@ pub fn quantize_tensor(tensor: Tensor<f32>, scale: Scale) -> Tensor<i32> {
 /// mask (`attention_score + mask_value`) drives the centered logit
 /// `z = max_k − (score + M) ≥ C·2^scale > z_bound`, so masked positions
 /// saturate to softmax weight 0 unconditionally — bit-identical to the old
-/// `-2^30` sentinel, for any in-row logits. Staying just one unit above the
-/// cutoff also keeps the masked sat_diff small enough to fit in the narrowed
-/// softmax range check with headroom for the in-row score spread that additive
-/// masks carry into sat_diff.
+/// `-2^30` sentinel, for any in-row logits.
 pub fn mask_sentinel_magnitude(scale: Scale) -> f64 {
     ((scale as f64 + 1.0) * std::f64::consts::LN_2).ceil() + 1.0
 }
@@ -151,14 +148,7 @@ pub fn quantize_float(float: f64, scale: Scale) -> i32 {
             // -(max_value/2) ≈ -2^30: mask_sentinel_magnitude(scale) stays
             // above the softmax exp round-to-zero cutoff (z_bound/2^scale ≈
             // (scale+1)·ln2, e.g. 6.75 at scale 8 → C=8), so masked positions
-            // still saturate to weight 0 bit-exactly, while the masked
-            // sat_diff fits in sat_diff_rc_bits(log_scale) = log_scale + 8
-            // (vs 30 bits), dropping the SatDiff one-hot commitment from d=8
-            // to d=4 (scale 8) polynomials per softmax node. Headroom for the
-            // score term that additive masks (attention_score + mask_value)
-            // carry into sat_diff is preserved: measured 13773 < 2^16 on the
-            // whole-model GPT-2 e2e workload and 13825 < 2^16 on BGE-small,
-            // both at scale 8 on the fused-i64 arithmetic. The sentinel must stay
+            // still saturate to weight 0 bit-exactly. The sentinel must stay
             // i32-representable: C(scale)·2^scale first exceeds i32::MAX at
             // scale 27, where the exp cutoff z_bound itself leaves the i32
             // range — no valid sentinel exists there at all, so fail loudly
