--- jolt-atlas-core/src/onnx_proof/ops/softmax_last_axis_satclamp/mod.rs
@@ -1,1152 +1,0 @@
-//! Benchmark-only sibling of `softmax_last_axis` that replaces the `sat_diff` complementary-
-//! slackness sumcheck with a `ClampBoundedTable`-based lookup
-//! (`joltworks::lookup_tables::clamp::SoftmaxSatClampTable`, via `sat_clamp::SoftmaxSatClampOperands`).
-//!
-//! Reuses `exp_sum`/`max`/`recip_mult`/`exponentiation`/`rc` directly from the original
-//! `softmax_last_axis` module (unchanged) — only the saturating-clamp step differs. See
-//! `sat_clamp.rs`'s module doc for the relation being proven, and the plan file
-//! (`i-want-to-benchmark-whimsical-tiger.md`) for the full design rationale.
-//!
-//! For more on the overall claim flow, see: https://hackmd.io/@R5TO3fi7TlWQKnY6Ejr2eg/Byjucoqjbx
-
-use crate::{
-    onnx_proof::{
-        op_lookups::OpLookupProvider,
-        ops::{
-            softmax_last_axis::{
-                exp_sum::{ExpSumParams, ExpSumProver, ExpSumVerifier},
-                exponentiation::{
-                    mult::{MultParams, MultProver, MultVerifier},
-                    ExpDigit, ExpReadRafProvider,
-                },
-                max::{MaxIndicatorParams, MaxIndicatorProver, MaxIndicatorVerifier},
-                rc::{SoftmaxRCProvider, SoftmaxRaEncoding},
-                recip_mult::{RecipMultParams, RecipMultProver, RecipMultVerifier},
-                LookupTableData, VerifierLookupTableData,
-            },
-            OperatorProofTrait,
-        },
-        ProofId, ProofType, Prover, Verifier,
-    },
-    utils::opening_access::{AccOpeningAccessor, Target},
-};
-use joltworks::{
-    config::{OneHotConfig, OneHotParams},
-    lookup_tables::clamp::SoftmaxSatClampTable,
-    poly::opening_proof::VerifierOpeningAccumulator,
-};
-use std::collections::BTreeMap;
-
-use atlas_onnx_tracer::{
-    node::ComputationNode,
-    ops::{
-        softmax::{
-            generate_exp_lut_decomposed, softmax_last_axis_decomposed_padded_bound, softmax_z,
-            SoftmaxLastAxisTrace,
-        },
-        SoftmaxLastAxisSatClamp,
-    },
-    tensor::Tensor,
-    utils::quantize::scale_to_multiplier,
-};
-use common::{
-    consts::{MODEL_SCALE, XLEN},
-    CommittedPoly, VirtualPoly,
-};
-use joltworks::{
-    field::JoltField,
-    poly::{
-        multilinear_polynomial::{MultilinearPolynomial, PolynomialEvaluation},
-        opening_proof::OpeningPoint,
-    },
-    subprotocols::{
-        identity_range_check::{identity_rangecheck_prover, identity_rangecheck_verifier},
-        shout,
-        sumcheck::{BatchedSumcheck, SumcheckInstanceProof},
-        sumcheck_prover::SumcheckInstanceProver,
-        sumcheck_verifier::SumcheckInstanceVerifier,
-    },
-    transcripts::Transcript,
-    utils::{errors::ProofVerifyError, lookup_bits::LookupBits, math::Math},
-};
-
-/// The saturating-clamp lookup helper (`SoftmaxSatClampOperands`) and its `ClampSpec` table.
-mod sat_clamp;
-use sat_clamp::SoftmaxSatClampOperands;
-
-impl<F: JoltField, T: Transcript> OperatorProofTrait<F, T> for SoftmaxLastAxisSatClamp {
-    #[tracing::instrument(skip_all, name = "SoftmaxLastAxisSatClamp::prove")]
-    fn prove(
-        &self,
-        node: &ComputationNode,
-        prover: &mut Prover<F, T>,
-    ) -> Vec<(ProofId, SumcheckInstanceProof<F, T>)> {
-        debug_assert_eq!(
-            self.scale, MODEL_SCALE as i32,
-            "SoftmaxLastAxisSatClamp only supports the currently-compiled MODEL_SCALE={MODEL_SCALE}"
-        );
-        let scale = scale_to_multiplier(self.scale) as i32;
-        let softmax_input = prover.trace.operand_tensors(node)[0];
-        let trace = softmax_last_axis_decomposed_padded_bound(softmax_input, scale).1;
-        SoftmaxLastAxisSatClampProver::new(node, trace, scale).prove(prover)
-    }
-
-    #[tracing::instrument(skip_all, name = "SoftmaxLastAxisSatClamp::verify")]
-    fn verify(
-        &self,
-        node: &ComputationNode,
-        verifier: &mut Verifier<'_, F, T>,
-    ) -> Result<(), ProofVerifyError> {
-        debug_assert_eq!(
-            self.scale, MODEL_SCALE as i32,
-            "SoftmaxLastAxisSatClamp only supports the currently-compiled MODEL_SCALE={MODEL_SCALE}"
-        );
-        let scale_bits = verifier.preprocessing.scale();
-        let scale = scale_to_multiplier(self.scale) as i32;
-        let mut sm = SoftmaxLastAxisSatClampVerifier::new(
-            node,
-            scale,
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-        );
-        sm.verify(
-            &mut verifier.accumulator,
-            &mut verifier.transcript,
-            verifier.proofs,
-            scale_bits,
-        )
-    }
-
-    fn get_committed_polynomials(&self, node: &ComputationNode) -> Vec<CommittedPoly> {
-        let log_scale = self.scale as usize;
-        let decomp = generate_exp_lut_decomposed(scale_to_multiplier(self.scale) as i32);
-        let log_hi = decomp.lut_hi.len().next_power_of_two().log_2();
-        let log_lo = decomp.lut_lo.len().next_power_of_two().log_2();
-        let idx = node.idx;
-
-        let mut polys = vec![];
-        for (log_k, ctor) in [
-            (
-                log_scale,
-                CommittedPoly::SoftmaxRemainderRaD as fn(usize, usize) -> _,
-            ),
-            (log_scale, CommittedPoly::SoftmaxExpRemainderRaD),
-            (XLEN, CommittedPoly::SoftmaxSatClampRaD),
-            (log_hi, CommittedPoly::SoftmaxZHiRaD),
-            (log_lo, CommittedPoly::SoftmaxZLoRaD),
-        ] {
-            let d =
-                OneHotParams::from_config_and_log_K(&OneHotConfig::default(), log_k).instruction_d;
-            polys.extend((0..d).map(|i| ctor(idx, i)));
-        }
-        polys
-    }
-}
-
-pub(crate) struct SoftmaxLastAxisSatClampProver {
-    pub(crate) computation_node: ComputationNode,
-    pub(crate) scale: i32,
-    pub(crate) F_N: [usize; 2],
-    pub(crate) trace: SoftmaxLastAxisTrace,
-    /// Sat-clamp lookup indices computed in stage 3, consumed by stage 4's one-hot checks.
-    sat_clamp_indices: Option<Vec<usize>>,
-}
-
-impl SoftmaxLastAxisSatClampProver {
-    #[inline]
-    fn idx(&self) -> usize {
-        self.computation_node.idx
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::prove", skip_all)]
-    /// Run the full proving pipeline, returning all stage proofs.
-    fn prove<F: JoltField, T: Transcript>(
-        mut self,
-        prover: &mut Prover<F, T>,
-    ) -> Vec<(ProofId, SumcheckInstanceProof<F, T>)> {
-        let scale_bits = prover.preprocessing.scale();
-
-        debug_assert_eq!(
-            self.scale.ilog2() as i32,
-            scale_bits,
-            "softmax operator scale (log2 {}) disagrees with preprocessing scale {scale_bits}",
-            self.scale.ilog2(),
-        );
-
-        // ── Pre-compute index/lookup data from trace vectors ────────────
-
-        // R-derived data (consumed by stage1 RC + stage2 one-hot)
-        let r_lookup_bits = to_lookup_bits(&self.trace.R, scale_bits as usize);
-        let r_indices = to_indices(&self.trace.R);
-
-        // r_exp-derived data (consumed by stage2 RC + stage3 one-hot)
-        let r_exp_lookup_bits =
-            to_lookup_bits(&self.trace.decomposed_exp.r_exp, scale_bits as usize);
-        let r_exp_indices = to_indices(&self.trace.decomposed_exp.r_exp);
-
-        // z_hi/z_lo indices (consumed by stage3 + stage4)
-        let z_hi_indices = to_indices(&self.trace.decomposed_exp.z_hi);
-        let z_lo_indices = to_indices(&self.trace.decomposed_exp.z_lo);
-
-        // Padded lookup tables (consumed by stage3 Shout + stage4 one-hot)
-        let base = self.trace.decomposed_exp.lut.base as u64;
-        let mut table_hi = std::mem::take(&mut self.trace.decomposed_exp.lut.lut_hi);
-        let z_bound_minus_1 = (table_hi.len() as u64) * base - 1;
-        pad_to_power_of_two(&mut table_hi);
-        let mut table_lo = std::mem::take(&mut self.trace.decomposed_exp.lut.lut_lo);
-        pad_to_power_of_two(&mut table_lo);
-
-        let lut_data = LookupTableData {
-            table_hi,
-            table_lo,
-            z_hi_indices,
-            z_lo_indices,
-            base,
-            z_bound_minus_1,
-        };
-
-        // z = max_k - x, the sat-clamp lookup's witness. Computed here (not inside
-        // `build_stage3_instances`) because stage 2 takes ownership of `self.trace.x` (for
-        // `max_indicator_prover`) — by the time stage 3 runs, `self.trace.x` is empty.
-        let last_dim = self.F_N[1];
-        let z = softmax_z(&self.trace.x, &self.trace.max_k, last_dim);
-        let z_tensor = Tensor::new(Some(&z), &[z.len()])
-            .expect("softmax_z tensor construction")
-            .padded_next_power_of_two()
-            .map(|v| v as i64);
-
-        // ── Pipeline ────────────────────────────────────────────────────
-
-        self.send_auxiliary_vectors(prover);
-        self.cache_exp_sum(prover);
-        self.cache_R(prover);
-        let stage_1_proof = self.stage1(prover, r_lookup_bits);
-
-        self.cache_r_exp(prover);
-        let stage_2_proof = self.stage2(prover, r_exp_lookup_bits, &r_indices);
-
-        // z_hi/z_lo need to be cached fresh at r2 (established by stage 2's exp_mult instance)
-        // now that sat_diff's interactive binding no longer produces these claims.
-        self.cache_z_hi_lo(prover);
-
-        let stage_3_proof = self.stage3(prover, &lut_data, &r_exp_indices, z_tensor);
-
-        let stage_4_proof = self.stage4(prover, &lut_data);
-
-        vec![
-            (ProofId(self.idx(), ProofType::SoftmaxStage1), stage_1_proof),
-            (ProofId(self.idx(), ProofType::SoftmaxStage2), stage_2_proof),
-            (ProofId(self.idx(), ProofType::SoftmaxStage3), stage_3_proof),
-            (ProofId(self.idx(), ProofType::SoftmaxStage4), stage_4_proof),
-        ]
-    }
-}
-
-/// Verifier for the sat-clamp softmax variant. Mirror of [`SoftmaxLastAxisSatClampProver`].
-pub(crate) struct SoftmaxLastAxisSatClampVerifier {
-    computation_node: ComputationNode,
-    scale: i32,
-    F_N: [usize; 2],
-    exp_sum: Vec<i32>,
-    max_k: Vec<i32>,
-    argmax_k: Vec<usize>,
-}
-
-impl SoftmaxLastAxisSatClampVerifier {
-    #[inline]
-    pub(crate) fn idx(&self) -> usize {
-        self.computation_node.idx
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::verify", skip_all)]
-    pub(crate) fn verify<F: JoltField, T: Transcript>(
-        &mut self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-        proofs: &BTreeMap<ProofId, SumcheckInstanceProof<F, T>>,
-        scale_bits: i32,
-    ) -> Result<(), ProofVerifyError> {
-        self.cache_exp_sum(accumulator, transcript)?;
-        self.cache_R(accumulator, transcript);
-        self.run_stage(
-            ProofType::SoftmaxStage1,
-            self.build_stage1_verifiers(accumulator, transcript, scale_bits)?,
-            proofs,
-            accumulator,
-            transcript,
-        )?;
-
-        let lut = VerifierLookupTableData::new(self.scale);
-
-        self.cache_r_exp(accumulator, transcript);
-        self.run_stage(
-            ProofType::SoftmaxStage2,
-            self.build_stage2_verifiers(accumulator, transcript, scale_bits),
-            proofs,
-            accumulator,
-            transcript,
-        )?;
-
-        self.cache_z_hi_lo(accumulator, transcript);
-
-        self.run_stage(
-            ProofType::SoftmaxStage3,
-            self.build_stage3_verifiers(accumulator, transcript, &lut),
-            proofs,
-            accumulator,
-            transcript,
-        )?;
-
-        self.operand_link(accumulator)?;
-
-        self.run_stage(
-            ProofType::SoftmaxStage4,
-            self.build_stage4_verifiers(accumulator, transcript, &lut),
-            proofs,
-            accumulator,
-            transcript,
-        )?;
-        Ok(())
-    }
-
-    /// Drive a single batched-sumcheck stage given pre-built verifier instances.
-    fn run_stage<F: JoltField, T: Transcript>(
-        &self,
-        proof_type: ProofType,
-        instances: Vec<Box<dyn SumcheckInstanceVerifier<F, T>>>,
-        proofs: &BTreeMap<ProofId, SumcheckInstanceProof<F, T>>,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-    ) -> Result<(), ProofVerifyError> {
-        let proof = proofs
-            .get(&ProofId(self.idx(), proof_type))
-            .ok_or(ProofVerifyError::MissingProof(self.idx()))?;
-        let refs: Vec<&dyn SumcheckInstanceVerifier<F, T>> =
-            instances.iter().map(|b| b.as_ref()).collect();
-        BatchedSumcheck::verify(proof, refs, accumulator, transcript)?;
-        Ok(())
-    }
-}
-
-// ---------------------------------------------------------------------------
-// SoftmaxLastAxisSatClampProver Methods
-// ---------------------------------------------------------------------------
-
-impl SoftmaxLastAxisSatClampProver {
-    pub(crate) fn new(node: &ComputationNode, trace: SoftmaxLastAxisTrace, scale: i32) -> Self {
-        let (&n, leading_dims) = node
-            .output_dims
-            .split_last()
-            .expect("softmax node must have at least one output dimension");
-        let f = leading_dims.iter().product::<usize>();
-        Self {
-            computation_node: node.clone(),
-            scale,
-            F_N: [f, n],
-            trace,
-            sat_clamp_indices: None,
-        }
-    }
-
-    /// Send auxiliary vectors (max_k, exp_sum_q, argmax_k) to the transcript. Identical to
-    /// `softmax_last_axis::SoftmaxLastAxisProver::send_auxiliary_vectors`.
-    #[tracing::instrument(
-        name = "SoftmaxLastAxisSatClampProver::send_auxiliary_vectors",
-        skip_all
-    )]
-    pub(crate) fn send_auxiliary_vectors<F: JoltField, T: Transcript>(
-        &self,
-        prover: &mut Prover<F, T>,
-    ) {
-        let [f, _] = self.F_N;
-        let mut provider = AccOpeningAccessor::new(&mut prover.accumulator, &self.computation_node)
-            .into_provider(&mut prover.transcript, OpeningPoint::default());
-        for k in 0..f {
-            provider.append_advice(
-                |idx| VirtualPoly::SoftmaxSumOutput(idx, k),
-                F::from_u32(self.trace.exp_sum_q[k] as u32),
-            );
-            provider.append_advice(
-                |idx| VirtualPoly::SoftmaxMaxOutput(idx, k),
-                F::from_u32(self.trace.max_k[k] as u32),
-            );
-            provider.append_advice(
-                |idx| VirtualPoly::SoftmaxMaxIndex(idx, k),
-                F::from_u32(self.trace.argmax_k[k] as u32),
-            );
-        }
-    }
-
-    /// Cache `exp_sum_q(r0_lead)` as a virtual polynomial opening.
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::cache_exp_sum", skip_all)]
-    pub(crate) fn cache_exp_sum<F: JoltField, T: Transcript>(&self, prover: &mut Prover<F, T>) {
-        let accessor = AccOpeningAccessor::new(&mut prover.accumulator, &self.computation_node);
-        let r0 = accessor.get_reduced_opening().0;
-        let log_f = self.F_N[0].log_2();
-        let r_lead = r0.split_at(log_f).0;
-        let eval = MultilinearPolynomial::from(self.trace.exp_sum_q.clone()).evaluate(&r_lead.r);
-        let mut provider = accessor.into_provider(&mut prover.transcript, r_lead);
-        provider.append_advice(VirtualPoly::SoftmaxExpSum, eval);
-    }
-
-    /// Cache the R (remainder) polynomial to the accumulator.
-    pub(crate) fn cache_R<F: JoltField, T: Transcript>(&self, prover: &mut Prover<F, T>) {
-        let accessor = AccOpeningAccessor::new(&mut prover.accumulator, &self.computation_node);
-        let r0 = accessor.get_reduced_opening().0;
-        let eval = MultilinearPolynomial::from(self.trace.R.clone()).evaluate(&r0.r);
-        let mut provider = accessor.into_provider(&mut prover.transcript, r0);
-        provider.append_advice(VirtualPoly::SoftmaxRecipMultRemainder, eval);
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::stage1", skip_all)]
-    fn stage1<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        r_lookup_bits: Vec<LookupBits>,
-    ) -> SumcheckInstanceProof<F, T> {
-        let mut instances = self.build_stage1_instances(prover, r_lookup_bits);
-        run_batched_prove(&mut instances, prover)
-    }
-
-    pub(crate) fn build_stage1_instances<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        r_lookup_bits: Vec<LookupBits>,
-    ) -> Vec<Box<dyn SumcheckInstanceProver<F, T>>> {
-        #[cfg_attr(not(feature = "zk"), allow(unused_mut))]
-        let mut recip_mult_params = RecipMultParams::new(
-            self.computation_node.clone(),
-            self.scale,
-            self.F_N,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-        #[cfg(feature = "zk")]
-        {
-            recip_mult_params.inv_sum_evals =
-                self.trace.inv_sum.iter().map(|&x| F::from_i32(x)).collect();
-        }
-        let exp_q_for_sum = self.trace.exp_q.clone();
-        let recip_mult_prover = RecipMultProver::initialize(
-            std::mem::take(&mut self.trace.exp_q),
-            std::mem::take(&mut self.trace.inv_sum),
-            recip_mult_params,
-        );
-
-        let exp_sum_params = ExpSumParams::new(
-            self.computation_node.clone(),
-            self.F_N,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-        let exp_sum_prover = ExpSumProver::initialize(exp_q_for_sum, exp_sum_params);
-
-        let provider = SoftmaxRCProvider::remainder(
-            self.computation_node.clone(),
-            prover.preprocessing.scale(),
-        );
-        let rc_R_prover =
-            identity_rangecheck_prover(&provider, r_lookup_bits, &mut prover.accumulator);
-
-        vec![
-            Box::new(recip_mult_prover),
-            Box::new(exp_sum_prover),
-            Box::new(rc_R_prover),
-        ]
-    }
-
-    /// Cache the r_exp polynomial to the accumulator.
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::cache_r_exp", skip_all)]
-    pub(crate) fn cache_r_exp<F: JoltField, T: Transcript>(&self, prover: &mut Prover<F, T>) {
-        let accessor = AccOpeningAccessor::new(&mut prover.accumulator, &self.computation_node);
-        let r1 = accessor.get_advice(VirtualPoly::SoftmaxExpQ).0;
-        let eval =
-            MultilinearPolynomial::from(self.trace.decomposed_exp.r_exp.clone()).evaluate(&r1.r);
-        let mut provider = accessor.into_provider(&mut prover.transcript, r1);
-        provider.append_advice(VirtualPoly::SoftmaxExpRemainder, eval);
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::stage2", skip_all)]
-    fn stage2<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        r_exp_lookup_bits: Vec<LookupBits>,
-        r_indices: &[usize],
-    ) -> SumcheckInstanceProof<F, T> {
-        let mut instances = self.build_stage2_instances(prover, r_exp_lookup_bits, r_indices);
-        run_batched_prove(&mut instances, prover)
-    }
-
-    /// Build stage 2 sumcheck instances (no `sat_diff` here — the saturating clamp is a lookup
-    /// now, folded into stage 3/4 instead).
-    pub(crate) fn build_stage2_instances<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        r_exp_lookup_bits: Vec<LookupBits>,
-        r_indices: &[usize],
-    ) -> Vec<Box<dyn SumcheckInstanceProver<F, T>>> {
-        let [f, _n] = self.F_N;
-        let log_f = f.log_2();
-
-        let accessor = AccOpeningAccessor::new(&prover.accumulator, &self.computation_node);
-        let r1 = accessor.get_advice(VirtualPoly::SoftmaxExpQ).0;
-        let r1_k = r1.split_at(log_f).0;
-
-        let max_k_eval = MultilinearPolynomial::from(self.trace.max_k.clone()).evaluate(&r1_k.r);
-        let max_indicator_params = MaxIndicatorParams::new(
-            self.computation_node.clone(),
-            self.F_N,
-            std::mem::take(&mut self.trace.argmax_k),
-            max_k_eval,
-            &prover.accumulator,
-        );
-        let max_indicator_prover =
-            MaxIndicatorProver::initialize(std::mem::take(&mut self.trace.x), max_indicator_params);
-
-        let exp_mult_params = MultParams::new(
-            self.computation_node.clone(),
-            self.scale,
-            &prover.accumulator,
-        );
-        let exp_mult_prover = MultProver::initialize(
-            std::mem::take(&mut self.trace.decomposed_exp.exp_hi),
-            std::mem::take(&mut self.trace.decomposed_exp.exp_lo),
-            exp_mult_params,
-        );
-
-        let provider = SoftmaxRCProvider::exp_remainder(
-            self.computation_node.clone(),
-            prover.preprocessing.scale(),
-        );
-        let exp_r_rc_prover =
-            identity_rangecheck_prover(&provider, r_exp_lookup_bits, &mut prover.accumulator);
-
-        let encoding = SoftmaxRaEncoding::remainder(self.idx(), prover.preprocessing.scale());
-        let [R_ra_prover, R_hw_prover, R_bool_prover] = shout::ra_onehot_provers(
-            &encoding,
-            r_indices,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        vec![
-            Box::new(exp_mult_prover),
-            Box::new(max_indicator_prover),
-            Box::new(exp_r_rc_prover),
-            R_ra_prover,
-            R_hw_prover,
-            R_bool_prover,
-        ]
-    }
-
-    /// Cache `z_hi`/`z_lo` fresh at `r2` (mirrors `cache_r_exp`) — `r2` is `SoftmaxExpHi`'s
-    /// opening point, established by stage 2's `exp_mult_prover`, independent of `sat_diff`.
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::cache_z_hi_lo", skip_all)]
-    pub(crate) fn cache_z_hi_lo<F: JoltField, T: Transcript>(&self, prover: &mut Prover<F, T>) {
-        let accessor = AccOpeningAccessor::new(&mut prover.accumulator, &self.computation_node);
-        let r2 = accessor.get_advice(VirtualPoly::SoftmaxExpHi).0;
-        let z_hi_eval =
-            MultilinearPolynomial::from(self.trace.decomposed_exp.z_hi.clone()).evaluate(&r2.r);
-        let z_lo_eval =
-            MultilinearPolynomial::from(self.trace.decomposed_exp.z_lo.clone()).evaluate(&r2.r);
-        let mut provider = accessor.into_provider(&mut prover.transcript, r2);
-        provider.append_advice(VirtualPoly::SoftmaxZHi, z_hi_eval);
-        provider.append_advice(VirtualPoly::SoftmaxZLo, z_lo_eval);
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampProver::stage3", skip_all)]
-    fn stage3<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        lut_data: &LookupTableData,
-        r_exp_indices: &[usize],
-        z_tensor: Tensor<i64>,
-    ) -> SumcheckInstanceProof<F, T> {
-        let mut instances = self.build_stage3_instances(prover, lut_data, r_exp_indices, z_tensor);
-        run_batched_prove(&mut instances, prover)
-    }
-
-    /// Build stage 3 sumcheck instances: exp-digit lookups (unchanged) plus the new saturating-
-    /// clamp lookup's execution proof (folded directly into this batch, replacing the old
-    /// `sat_diff_rc_prover`). `z_tensor` is computed by the caller (`prove`), since by this point
-    /// `self.trace.x` has already been consumed by stage 2's `max_indicator_prover`.
-    pub(crate) fn build_stage3_instances<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        lut_data: &LookupTableData,
-        r_exp_indices: &[usize],
-        z_tensor: Tensor<i64>,
-    ) -> Vec<Box<dyn SumcheckInstanceProver<F, T>>> {
-        let hi_provider = ExpReadRafProvider {
-            node: self.computation_node.clone(),
-            table_size: lut_data.table_hi.len(),
-            digit: ExpDigit::Hi,
-        };
-        let hi_prover = shout::read_raf_prover(
-            &hi_provider,
-            &lut_data.z_hi_indices,
-            &lut_data.table_hi,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        let lo_provider = ExpReadRafProvider {
-            node: self.computation_node.clone(),
-            table_size: lut_data.table_lo.len(),
-            digit: ExpDigit::Lo,
-        };
-        let lo_prover = shout::read_raf_prover(
-            &lo_provider,
-            &lut_data.z_lo_indices,
-            &lut_data.table_lo,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        let sat_clamp_provider = OpLookupProvider::with_helper(
-            self.computation_node.clone(),
-            SoftmaxSatClampOperands::new(z_tensor),
-        );
-        let (sat_clamp_prover, sat_clamp_indices) = sat_clamp_provider
-            .read_raf_prove::<F, T, SoftmaxSatClampTable<XLEN>, XLEN>(
-                &prover.trace,
-                &mut prover.accumulator,
-                &mut prover.transcript,
-            );
-
-        // Stash the lookup indices for stage 4's one-hot checks.
-        self.sat_clamp_indices = Some(sat_clamp_indices);
-
-        let encoding = SoftmaxRaEncoding::exp_remainder(self.idx(), prover.preprocessing.scale());
-        let [exp_r_ra_prover, exp_r_hw_prover, exp_r_bool_prover] = shout::ra_onehot_provers(
-            &encoding,
-            r_exp_indices,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        vec![
-            hi_prover,
-            lo_prover,
-            Box::new(sat_clamp_prover),
-            exp_r_ra_prover,
-            exp_r_hw_prover,
-            exp_r_bool_prover,
-        ]
-    }
-
-    fn stage4<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        lut_data: &LookupTableData,
-    ) -> SumcheckInstanceProof<F, T> {
-        let mut instances = self.build_stage4_instances(prover, lut_data);
-        run_batched_prove(&mut instances, prover)
-    }
-
-    /// Build stage 4 sumcheck instances: exp-digit one-hot checks (unchanged) plus the new
-    /// saturating-clamp lookup's one-hot checks (replacing the old `sat_diff` triple).
-    pub(crate) fn build_stage4_instances<F: JoltField, T: Transcript>(
-        &mut self,
-        prover: &mut Prover<F, T>,
-        lut_data: &LookupTableData,
-    ) -> Vec<Box<dyn SumcheckInstanceProver<F, T>>> {
-        let encoding = SoftmaxRaEncoding::exp_hi(self.idx(), lut_data.table_hi.len().log_2());
-        let [hi_ra_prover, hi_hw_prover, hi_bool_prover] = shout::ra_onehot_provers(
-            &encoding,
-            &lut_data.z_hi_indices,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        let encoding = SoftmaxRaEncoding::exp_lo(self.idx(), lut_data.table_lo.len().log_2());
-        let [lo_ra_prover, lo_hw_prover, lo_bool_prover] = shout::ra_onehot_provers(
-            &encoding,
-            &lut_data.z_lo_indices,
-            &prover.accumulator,
-            &mut prover.transcript,
-        );
-
-        let sat_clamp_indices = self
-            .sat_clamp_indices
-            .take()
-            .expect("stage 3 must run before stage 4");
-        let sat_clamp_provider: OpLookupProvider<SoftmaxSatClampOperands> =
-            OpLookupProvider::new(self.computation_node.clone());
-        let encoding = sat_clamp_provider.encoding();
-        let [sat_clamp_ra_prover, sat_clamp_hw_prover, sat_clamp_bool_prover] =
-            shout::ra_onehot_provers(
-                &encoding,
-                &sat_clamp_indices,
-                &prover.accumulator,
-                &mut prover.transcript,
-            );
-
-        vec![
-            hi_ra_prover,
-            hi_hw_prover,
-            hi_bool_prover,
-            lo_ra_prover,
-            lo_hw_prover,
-            lo_bool_prover,
-            sat_clamp_ra_prover,
-            sat_clamp_hw_prover,
-            sat_clamp_bool_prover,
-        ]
-    }
-}
-
-// ---------------------------------------------------------------------------
-// SoftmaxLastAxisSatClampVerifier methods
-// ---------------------------------------------------------------------------
-
-impl SoftmaxLastAxisSatClampVerifier {
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::new", skip_all)]
-    pub(crate) fn new<F: JoltField, T: Transcript>(
-        node: &ComputationNode,
-        scale: i32,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-    ) -> Self {
-        let (&n, leading_dims) = node
-            .output_dims
-            .split_last()
-            .expect("softmax node must have at least one output dimension");
-        let f = leading_dims.iter().product::<usize>();
-
-        let accessor = AccOpeningAccessor::new(&mut *accumulator, node);
-        let mut provider = accessor.into_provider(&mut *transcript, OpeningPoint::default());
-        for k in 0..f {
-            provider.append_advice(|idx| VirtualPoly::SoftmaxSumOutput(idx, k));
-            provider.append_advice(|idx| VirtualPoly::SoftmaxMaxOutput(idx, k));
-            provider.append_advice(|idx| VirtualPoly::SoftmaxMaxIndex(idx, k));
-        }
-
-        fn read_aux_scalars<F: JoltField>(
-            accessor: &AccOpeningAccessor<'_, F, VerifierOpeningAccumulator<F>>,
-            f: usize,
-            vp_fn: fn(usize, usize) -> VirtualPoly,
-        ) -> Vec<u64> {
-            (0..f)
-                .map(|k| {
-                    accessor
-                        .get_advice(|idx| vp_fn(idx, k))
-                        .1
-                        .to_u64()
-                        .expect("auxiliary scalar should fit within 64 bits")
-                })
-                .collect()
-        }
-
-        let accessor = AccOpeningAccessor::new(&*accumulator, node);
-
-        let exp_sum = read_aux_scalars(&accessor, f, VirtualPoly::SoftmaxSumOutput)
-            .into_iter()
-            .map(|v| v as i32)
-            .collect();
-        let max_k = read_aux_scalars(&accessor, f, VirtualPoly::SoftmaxMaxOutput)
-            .into_iter()
-            .map(|v| v as i32)
-            .collect();
-        let argmax_k = read_aux_scalars(&accessor, f, VirtualPoly::SoftmaxMaxIndex)
-            .into_iter()
-            .map(|v| v as usize)
-            .collect();
-
-        Self {
-            computation_node: node.clone(),
-            scale,
-            F_N: [f, n],
-            exp_sum,
-            max_k,
-            argmax_k,
-        }
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::inv_sum_evals", skip_all)]
-    pub(crate) fn inv_sum_evals<F: JoltField>(
-        &self,
-        accumulator: &VerifierOpeningAccumulator<F>,
-    ) -> Result<Vec<F>, ProofVerifyError> {
-        let [f, _] = self.F_N;
-        let s_squared = (self.scale as i64) * (self.scale as i64);
-        let accessor = AccOpeningAccessor::new(accumulator, &self.computation_node);
-        (0..f)
-            .map(|k| {
-                let exp_sum_q_k = accessor
-                    .get_advice(|idx| VirtualPoly::SoftmaxSumOutput(idx, k))
-                    .1;
-                let exp_sum_q_int = exp_sum_q_k
-                    .to_u64()
-                    .expect("exp_sum_q[k] should fit in u64")
-                    as i64;
-                if exp_sum_q_int == 0 {
-                    return Err(ProofVerifyError::InvalidOpeningProof(format!(
-                        "exp_sum_q[{k}] is zero, cannot compute inv_sum"
-                    )));
-                }
-                Ok(F::from_i32((s_squared / exp_sum_q_int) as i32))
-            })
-            .collect()
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::cache_exp_sum", skip_all)]
-    pub(crate) fn cache_exp_sum<F: JoltField, T: Transcript>(
-        &mut self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-    ) -> Result<(), ProofVerifyError> {
-        let accessor = AccOpeningAccessor::new(accumulator, &self.computation_node);
-        let r0 = accessor.get_reduced_opening().0;
-        let log_f = self.F_N[0].log_2();
-        let r_lead = r0.split_at(log_f).0;
-        let mut provider = accessor.into_provider(transcript, r_lead.clone());
-        provider.append_advice(VirtualPoly::SoftmaxExpSum);
-        let exp_sum_eval =
-            MultilinearPolynomial::from(std::mem::take(&mut self.exp_sum)).evaluate(&r_lead.r);
-        let claimed_exp_sum_eval = provider.get_advice(VirtualPoly::SoftmaxExpSum).1;
-        if exp_sum_eval != claimed_exp_sum_eval {
-            return Err(ProofVerifyError::InvalidOpeningProof(
-                "exp_sum evaluation mismatch".to_string(),
-            ));
-        }
-        Ok(())
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::cache_R", skip_all)]
-    pub(crate) fn cache_R<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-    ) {
-        let accessor = AccOpeningAccessor::new(accumulator, &self.computation_node);
-        let r = accessor.get_reduced_opening().0;
-        let mut provider = accessor.into_provider(transcript, r);
-        provider.append_advice(VirtualPoly::SoftmaxRecipMultRemainder);
-    }
-
-    #[tracing::instrument(
-        name = "SoftmaxLastAxisSatClampVerifier::build_stage1_verifiers",
-        skip_all
-    )]
-    pub(crate) fn build_stage1_verifiers<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-        scale_bits: i32,
-    ) -> Result<Vec<Box<dyn SumcheckInstanceVerifier<F, T>>>, ProofVerifyError> {
-        let inv_sum = self.inv_sum_evals(&*accumulator)?;
-        let recip_mult_verifier = RecipMultVerifier::new(
-            self.computation_node.clone(),
-            self.scale,
-            self.F_N,
-            inv_sum,
-            &*accumulator,
-            transcript,
-        );
-        let exp_sum_verifier = ExpSumVerifier::new(
-            self.computation_node.clone(),
-            self.F_N,
-            &*accumulator,
-            transcript,
-        );
-        let rc_provider = SoftmaxRCProvider::remainder(self.computation_node.clone(), scale_bits);
-        let rc_R_verifier = identity_rangecheck_verifier(&rc_provider, accumulator);
-        Ok(vec![
-            Box::new(recip_mult_verifier),
-            Box::new(exp_sum_verifier),
-            Box::new(rc_R_verifier),
-        ])
-    }
-
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::cache_r_exp", skip_all)]
-    pub(crate) fn cache_r_exp<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-    ) {
-        let accessor = AccOpeningAccessor::new(accumulator, &self.computation_node);
-        let r = accessor.get_advice(VirtualPoly::SoftmaxExpQ).0;
-        let mut provider = accessor.into_provider(transcript, r);
-        provider.append_advice(VirtualPoly::SoftmaxExpRemainder);
-    }
-
-    #[tracing::instrument(
-        name = "SoftmaxLastAxisSatClampVerifier::build_stage2_verifiers",
-        skip_all
-    )]
-    pub(crate) fn build_stage2_verifiers<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-        scale_bits: i32,
-    ) -> Vec<Box<dyn SumcheckInstanceVerifier<F, T>>> {
-        let [f, _n] = self.F_N;
-        let log_f = f.log_2();
-
-        let accessor = AccOpeningAccessor::new(&*accumulator, &self.computation_node);
-        let r1 = accessor.get_advice(VirtualPoly::SoftmaxExpQ).0;
-        let r1_k = r1.split_at(log_f).0;
-
-        let max_k_eval = MultilinearPolynomial::from(self.max_k.clone()).evaluate(&r1_k.r);
-        let max_indicator_verifier = MaxIndicatorVerifier::new(
-            self.computation_node.clone(),
-            self.F_N,
-            self.argmax_k.clone(),
-            max_k_eval,
-            &*accumulator,
-        );
-
-        let exp_mult_verifier =
-            MultVerifier::new(self.computation_node.clone(), self.scale, &*accumulator);
-
-        let rc_provider =
-            SoftmaxRCProvider::exp_remainder(self.computation_node.clone(), scale_bits);
-        let exp_r_rc_verifier = identity_rangecheck_verifier(&rc_provider, accumulator);
-
-        let encoding = SoftmaxRaEncoding::remainder(self.idx(), scale_bits);
-        let [R_ra_verifier, R_hw_verifier, R_bool_verifier] =
-            shout::ra_onehot_verifiers(&encoding, &*accumulator, transcript);
-
-        vec![
-            Box::new(exp_mult_verifier),
-            Box::new(max_indicator_verifier),
-            Box::new(exp_r_rc_verifier),
-            R_ra_verifier,
-            R_hw_verifier,
-            R_bool_verifier,
-        ]
-    }
-
-    /// Verifier counterpart of [`SoftmaxLastAxisSatClampProver::cache_z_hi_lo`]: append the
-    /// `SoftmaxZHi`/`SoftmaxZLo` opening requests (claims loaded from the proof).
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::cache_z_hi_lo", skip_all)]
-    pub(crate) fn cache_z_hi_lo<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-    ) {
-        let accessor = AccOpeningAccessor::new(accumulator, &self.computation_node);
-        let r2 = accessor.get_advice(VirtualPoly::SoftmaxExpHi).0;
-        let mut provider = accessor.into_provider(transcript, r2);
-        provider.append_advice(VirtualPoly::SoftmaxZHi);
-        provider.append_advice(VirtualPoly::SoftmaxZLo);
-    }
-
-    #[tracing::instrument(
-        name = "SoftmaxLastAxisSatClampVerifier::build_stage3_verifiers",
-        skip_all
-    )]
-    pub(crate) fn build_stage3_verifiers<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &mut VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-        lut: &VerifierLookupTableData,
-    ) -> Vec<Box<dyn SumcheckInstanceVerifier<F, T>>> {
-        let provider = ExpReadRafProvider {
-            node: self.computation_node.clone(),
-            table_size: lut.table_hi.len(),
-            digit: ExpDigit::Hi,
-        };
-        let hi_verifier =
-            shout::read_raf_verifier(&provider, lut.table_hi.clone(), &*accumulator, transcript);
-
-        let provider = ExpReadRafProvider {
-            node: self.computation_node.clone(),
-            table_size: lut.table_lo.len(),
-            digit: ExpDigit::Lo,
-        };
-        let lo_verifier =
-            shout::read_raf_verifier(&provider, lut.table_lo.clone(), &*accumulator, transcript);
-
-        let sat_clamp_provider: OpLookupProvider<SoftmaxSatClampOperands> =
-            OpLookupProvider::new(self.computation_node.clone());
-        let sat_clamp_verifier = sat_clamp_provider
-            .read_raf_verify::<F, T, SoftmaxSatClampTable<XLEN>, XLEN>(accumulator, transcript);
-
-        let scale_bits = self.scale.ilog2() as i32;
-        let encoding = SoftmaxRaEncoding::exp_remainder(self.idx(), scale_bits);
-        let [r_exp_ra_verifier, r_exp_hw_verifier, r_exp_bool_verifier] =
-            shout::ra_onehot_verifiers(&encoding, &*accumulator, transcript);
-
-        vec![
-            hi_verifier,
-            lo_verifier,
-            Box::new(sat_clamp_verifier),
-            r_exp_ra_verifier,
-            r_exp_hw_verifier,
-            r_exp_bool_verifier,
-        ]
-    }
-
-    /// Operand link: derive `X(r2)` algebraically and route upstream.
-    ///
-    /// `X(r2) = max_k(r2_lead) − z(r2)`, where `z(r2)` is the sat-clamp lookup's witness claim
-    /// (`z = max_k - x`, already tied to `z_c = ClampTable[z]` and `z_hi*base + z_lo` by the
-    /// stage-3 lookup itself — see `sat_clamp.rs`).
-    #[tracing::instrument(name = "SoftmaxLastAxisSatClampVerifier::operand_link", skip_all)]
-    pub(crate) fn operand_link<F: JoltField>(
-        &self,
-        accumulator: &VerifierOpeningAccumulator<F>,
-    ) -> Result<(), ProofVerifyError> {
-        let [f, _n] = self.F_N;
-        let log_f = f.log_2();
-        let accessor = AccOpeningAccessor::new(accumulator, &self.computation_node);
-
-        let r2 = accessor.get_advice(VirtualPoly::SoftmaxExpHi).0;
-        let r2_lead = r2.split_at(log_f).0;
-
-        let max_k_eval = MultilinearPolynomial::from(self.max_k.clone()).evaluate(&r2_lead.r);
-
-        let z_eval = accessor.get_advice(VirtualPoly::SoftmaxSatClampWitness).1;
-
-        let x_r2 = max_k_eval - z_eval;
-
-        let prover_x_r2 = accessor.get_nodeio(Target::Input(0)).1;
-        if prover_x_r2 != x_r2 {
-            return Err(ProofVerifyError::InvalidOpeningProof(
-                "Operand link failed: prover's X(r2) does not match max_k - z".to_string(),
-            ));
-        }
-        Ok(())
-    }
-
-    #[tracing::instrument(
-        name = "SoftmaxLastAxisSatClampVerifier::build_stage4_verifiers",
-        skip_all
-    )]
-    pub(crate) fn build_stage4_verifiers<F: JoltField, T: Transcript>(
-        &self,
-        accumulator: &VerifierOpeningAccumulator<F>,
-        transcript: &mut T,
-        lut: &VerifierLookupTableData,
-    ) -> Vec<Box<dyn SumcheckInstanceVerifier<F, T>>> {
-        let encoding = SoftmaxRaEncoding::exp_hi(self.idx(), lut.table_hi.len().log_2());
-        let [hi_ra_verifier, hi_hw_verifier, hi_bool_verifier] =
-            shout::ra_onehot_verifiers(&encoding, accumulator, transcript);
-
-        let encoding = SoftmaxRaEncoding::exp_lo(self.idx(), lut.table_lo.len().log_2());
-        let [lo_ra_verifier, lo_hw_verifier, lo_bool_verifier] =
-            shout::ra_onehot_verifiers(&encoding, accumulator, transcript);
-
-        let sat_clamp_provider: OpLookupProvider<SoftmaxSatClampOperands> =
-            OpLookupProvider::new(self.computation_node.clone());
-        let encoding = sat_clamp_provider.encoding();
-        let [sat_clamp_ra_verifier, sat_clamp_hw_verifier, sat_clamp_bool_verifier] =
-            shout::ra_onehot_verifiers(&encoding, accumulator, transcript);
-
-        vec![
-            hi_ra_verifier,
-            hi_hw_verifier,
-            hi_bool_verifier,
-            lo_ra_verifier,
-            lo_hw_verifier,
-            lo_bool_verifier,
-            sat_clamp_ra_verifier,
-            sat_clamp_hw_verifier,
-            sat_clamp_bool_verifier,
-        ]
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Small helpers used across prover stages
-// ---------------------------------------------------------------------------
-
-/// Convert `i32` trace values to usize indices.
-pub(crate) fn to_indices(values: &[i32]) -> Vec<usize> {
-    values.iter().map(|&v| v as usize).collect()
-}
-
-/// Convert `i32` trace values to `LookupBits` with the given bit-width.
-pub(crate) fn to_lookup_bits(values: &[i32], bits: usize) -> Vec<LookupBits> {
-    values
-        .iter()
-        .map(|&v| LookupBits::new(v as u64, bits))
-        .collect()
-}
-
-/// Pad a vector in-place to the next power of two with zeros.
-pub(crate) fn pad_to_power_of_two(v: &mut Vec<i32>) {
-    v.resize(v.len().next_power_of_two(), 0);
-}
-
-/// Run `BatchedSumcheck::prove` with the standard boilerplate.
-fn run_batched_prove<F: JoltField, T: Transcript>(
-    instances: &mut [Box<dyn SumcheckInstanceProver<F, T>>],
-    prover: &mut Prover<F, T>,
-) -> SumcheckInstanceProof<F, T> {
-    BatchedSumcheck::prove(
-        instances.iter_mut().map(|v| &mut **v as _).collect(),
-        &mut prover.accumulator,
-        &mut prover.transcript,
-    )
-    .0
-}
-
-#[cfg(test)]
-mod tests {
-    use atlas_onnx_tracer::{
-        model::{test::ModelBuilder, Model},
-        tensor::Tensor,
-    };
-
-    use crate::onnx_proof::ops::test::unit_test_op;
-
-    fn softmax_last_axis_satclamp_model(input_shape: &[usize], scale: u32) -> Model {
-        let mut b = ModelBuilder::with_scale(scale);
-        let i = b.input(input_shape.to_vec());
-        let res = b.softmax_last_axis_satclamp(i);
-        b.mark_output(res);
-        b.build()
-    }
-
-    /// Helper: build a small softmax test at the given scale.
-    fn run_softmax_scale_test(scale: u32) {
-        let input_shape = vec![2, 8];
-        #[rustfmt::skip]
-        let data: Vec<i32> = vec![
-            10, 20, 30, 40, 50, 60, 5, 15,
-            25, 35, 45, 55, 8, 18, 28, 38,
-        ];
-        let input = Tensor::new(Some(&data), &input_shape).unwrap();
-        let model = softmax_last_axis_satclamp_model(&input_shape, scale);
-        unit_test_op(model, &[input]);
-    }
-
-    #[test]
-    fn test_softmax_satclamp_model_scale() {
-        run_softmax_scale_test(common::consts::MODEL_SCALE as u32);
-    }
-
-    /// GPT-2-sized causal-attention fixture (4 heads, 8x8 each), at whatever `MODEL_SCALE` is
-    /// currently compiled to — exercises a realistic shape/magnitude, not just the tiny fixture.
-    #[test]
-    fn test_softmax_satclamp_gpt2_sized() {
-        use atlas_onnx_tracer::utils::quantize::mask_sentinel_magnitude;
-        use rand::{rngs::StdRng, Rng, SeedableRng};
-
-        let scale = common::consts::MODEL_SCALE as i32;
-        let mask_c = mask_sentinel_magnitude(scale) as i32;
-        let mask_value = -(mask_c << scale);
-
-        let input_shape = vec![4, 8, 8];
-        let mut rng = StdRng::seed_from_u64(0x5a7c);
-        let mut data = vec![0i32; 4 * 8 * 8];
-        for head in 0..4 {
-            for row in 0..8 {
-                for col in 0..8 {
-                    let idx = head * 64 + row * 8 + col;
-                    data[idx] = if col > row {
-                        mask_value
-                    } else {
-                        rng.gen_range(-(1 << (scale + 2))..(1 << (scale + 2)))
-                    };
-                }
-            }
-        }
-        let input = Tensor::new(Some(&data), &input_shape).unwrap();
-        let model = softmax_last_axis_satclamp_model(&input_shape, scale as u32);
-        unit_test_op(model, &[input]);
-    }
-}
